sos_seva/requirements.txt
django==2.2.6
django-datatables-view==1.19.1
django-import-export==2.5.0
google-cloud-translate==3.1.0
langdetect==1.0.8
mysqlclient==2.0.3
pandas==1.1.5
nltk==3.6.2
scipy==1.6.3
searchtweets==1.7.6
sent2vec==0.2.0
torch==1.8.1
tweepy==3.10.0
twilio==6.38.0